gacp

+++ Cars/car_30June2018/nodes/network_node.py
@@ -298,12 +298,14 @@ while not rospy.is_shutdown():
     else:
         caffe_enter_timer.reset()
     
-    shutdown_time = 30
-    if state == 4 and time.time()-state_enter_time > shutdown_time-5:
-        print('!!! about to reboot from state 4 !!! ' + str(steer))
-    if state == 4 and time.time()-state_enter_time > shutdown_time:
-        print(d2s("Rebooting because in state 4 for",shutdown_time,"+ s"))
-        unix('sudo reboot')
+    if False:
+        shutdown_time = 30
+        if state == 4 and time.time()-state_enter_time > shutdown_time-5:
+            print('!!! about to reboot from state 4 !!! ' + str(steer))
+        if state == 4 and time.time()-state_enter_time > shutdown_time:
+            print(d2s("Rebooting because in state 4 for",shutdown_time,"+ s"))
+            unix('sudo reboot')
+
     if time_step.check():
         print(d2s("In state",state,"for",dp(time.time()-state_enter_time),"seconds, previous_state =",previous_state))
         time_step.reset()
